+++ back-end/server.js
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 const server = require("./app") // load up the web server
-const port = 3001 // the port to listen to for incoming requests
+const port = 3000// the port to listen to for incoming requests
 // call express's listen function to start listening to the port
 //const username = process.argv[2].split('=')[1]
 //console.log(`hi, ${username}`)
